# docs/openapi.yaml
openapi: 3.1.0
info:
  title: AgentAuth Learning Lab API
  version: 0.4.0-beta
  description: >-
    RFC-0150 compliant authorization framework API for learning, experimentation, and compliance validation.
    This comprehensive implementation provides endpoints for evaluation and demonstration of AgentAuth concepts,
    authorization pattern testing, and RFC-0150 compliance verification.
    
    **Learning Features:**
    - Interactive authorization pattern evaluation
    - RFC-0150 compliance testing and validation
    - Educational sandbox environments for experimentation
    - Comprehensive audit trails and transparency logs
    
    **Implementation Status:**
    - Core Authorization: ✅ Complete
    - Delegation & PoA: ✅ Complete  
    - Revocation & Lifecycle: ✅ Complete
    - Audit & Transparency: ✅ Complete
    - Cryptographic Standards: ✅ Complete
    - API & Integration: ✅ Complete
...
servers:
  - url: http://localhost:8080
    description: Local development server
paths:
...
  /api/v1/token/create:
    post:
      summary: Create access token
      tags: [validation]
      operationId: postTokenCreate
      description: |
        Issues a new access token with optional TTL and metadata.
        If GAUTH_USE_JWT_LIB=1, also returns a signed JWT in the response.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                ttl_seconds:
                  type: integer
                  description: Token time-to-live in seconds.
                meta:
                  type: object
                  description: Arbitrary metadata to attach to the token.
                nonce:
                  type: string
                  description: Optional client-provided nonce for replay protection.
      responses:
        '201':
          description: Token created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  token:
                    type: object
                    properties:
                      id: { type: string }
                      value: { type: string }
                      expires_at: { type: string, format: date-time }
                  jwt: { type: string, description: Signed JWT (if enabled) }
        '403':
          description: Capability denied
        '409':
          description: Nonce replay detected
        '500':
          description: Internal error
  /api/v1/token/validate:
    post:
      summary: Validate access token
      tags: [validation]
      operationId: postTokenValidate
      description: |
        Validates a token and returns its status and claims.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token]
              properties:
                token:
                  type: string
                  description: The token value or JWT to validate.
                token_id:
                  type: string
                  description: Alternative lookup by token ID.
      responses:
        '200':
          description: Validation result
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  valid: { type: boolean }
                  reason: { type: string }
                  claims: { type: object }
        '400':
          description: Invalid request
  /api/v1/token/revoke:
    post:
      summary: Revoke access token
      tags: [validation]
      operationId: postTokenRevoke
      description: |
        Revokes an active token, making it invalid for future use.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token_id]
              properties:
                token_id:
                  type: string
                  description: ID of the token to revoke.
                reason:
                  type: string
                  description: Optional revocation reason.
      responses:
        '200':
          description: Token revoked
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  revoked_at: { type: string, format: date-time }
        '404':
          description: Token not found
...